+++ Testing/integration/tests/Usings.cs
@@ -1,1 +1,2 @@
+global using Shouldly;
 global using Xunit;
